update changes add check_destination

+++ plugins/action/net_put.py
@@ -13,7 +13,7 @@ import uuid
 
 from ansible.errors import AnsibleError
 from ansible.module_utils._text import to_bytes, to_text
-from ansible.module_utils.connection import Connection, ConnectionError
+from ansible.module_utils.connection import Connection
 from ansible.module_utils.six.moves.urllib.parse import urlsplit
 from ansible.plugins.action import ActionBase
 from ansible.utils.display import Display
@@ -64,6 +64,11 @@ class ActionModule(ActionBase):
         if mode is None:
             mode = "binary"
 
+        # Check if the file is present in destination or not
+        check_destination = self._task.args.get("check_destination")
+        if check_destination is None:
+            check_destination = True
+
         if mode == "text":
             try:
                 self._handle_src_option(convert_data=False)
@@ -95,32 +100,38 @@ class ActionModule(ActionBase):
 
         if dest is None:
             dest = src_file_path_name
-        try:
-            changed = self._handle_existing_file(conn, output_file, dest, proto, sock_timeout)
-            if changed is False:
-                result["changed"] = changed
-                result["destination"] = dest
-                if mode == "text":
-                    # Cleanup tmp file expanded wih ansible vars
-                    os.remove(output_file)
-                return result
-        except Exception as exc:
-            result["msg"] = "Warning: %s idempotency check failed. Check dest" % exc
+
+        if check_destination:
+            try:
+                changed = self._handle_existing_file(
+                    conn, output_file, dest, proto, sock_timeout
+                )
+                if changed is False:
+                    result["changed"] = changed
+                    result["destination"] = dest
+                    if mode == "text":
+                        # Cleanup tmp file expanded wih ansible vars
+                        os.remove(output_file)
+                    return result
+            except Exception as exc:
+                result["msg"] = "Warning: %s idempotency check failed. Check dest" % exc
 
         try:
-            # libssh fails if file is not present in remote appliance
             conn.copy_file(
                 source=output_file,
                 destination=dest,
                 proto=proto,
                 timeout=sock_timeout,
             )
+            changed = True
         except Exception as exc:
             if to_text(exc) == "No response from server":
                 if network_os == "iosxr":
                     # IOSXR sometimes closes socket prematurely after completion
                     # of file transfer
-                    result["msg"] = "Warning: iosxr scp server pre close issue. Please check dest"
+                    result["msg"] = (
+                        "Warning: iosxr scp server pre close issue. Please check dest"
+                    )
             else:
                 result["failed"] = True
                 result["msg"] = "Exception received: %s" % exc
@@ -155,6 +166,7 @@ class ActionModule(ActionBase):
                 if os.path.exists(tmp_source_file):
                     os.remove(tmp_source_file)
                 return True
+
         try:
             with open(source, "r") as f:
                 new_content = f.read()
